various minor fixes

+++ cost/stats.php
@@ -63,9 +63,9 @@ dol_include_once('/ultimateimmo/class/immoowner.class.php');
 $langs->loadLangs(array("ultimateimmo@ultimateimmo", "other", "bills"));
 
 // Filter
-$year = GETPOST("year", 'int');
+$year = (int)GETPOST("year", 'int');
 if ($year == 0) {
-	$year_current = strftime("%Y", time());
+	$year_current = (int)strftime("%Y", time());
 	$year_start = $year_current;
 } else {
 	$year_current = $year;
